pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "planetarble"
version = "0.1.0"
description = "Reproducible pipeline for generating global PMTiles from open data sources."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "CC0-1.0" }
authors = [
  { name = "Planetarble Contributors" }
]
dependencies = [
  "PyYAML>=6.0"
]

[project.scripts]
planetarble = "planetarble.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]
